+++ docs/source/conf.py
@@ -13,10 +13,11 @@
 # documentation root, use os.path.abspath to make it absolute, like shown here.
 #
 import sys
+import os
 
 sys.path.insert(0, os.path.abspath('../../'))
 
-__version__ = '0.4.0'
+__version__ = '0.4.2'
 
 # -- Project information -----------------------------------------------------