+++ examples/python/matrix_set_operations.py
@@ -1,6 +1,6 @@
 # encoding: utf-8
 
-# (c) 2017-2018 Open Risk, all rights reserved
+# (c) 2017-2019 Open Risk, all rights reserved
 #
 # TransitionMatrix is licensed under the Apache 2.0 license a copy of which is included
 # in the source distribution of TransitionMatrix. This is notwithstanding any licenses of
@@ -18,7 +18,7 @@
 """
 
 import transitionMatrix as tm
-from datasets import Generic as T1
+from transitionMatrix.predefined import Generic as T1
 
 print("-- Lets seed the set with a 3x3 matrix")
 A = tm.TransitionMatrix(values=[[0.6, 0.2, 0.2], [0.2, 0.6, 0.2], [0.2, 0.2, 0.6]])