+++ run_examples.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 
-# (c) 2017-2018 Open Risk, all rights reserved
+# (c) 2017-2019 Open Risk, all rights reserved
 #
 # TransitionMatrix is licensed under the Apache 2.0 license a copy of which is included
 # in the source distribution of TransitionMatrix. This is notwithstanding any licenses of
@@ -23,7 +23,7 @@ examples_path = source_path + "examples/python/"
 filelist = ['adjust_nr_state', 'credit_curves',
             'empirical_transition_matrix', 'fix_multiperiod_matrix', 'generate_synthetic_data',
             'generate_visuals', 'matrix_from_cohort_data', 'matrix_from_duration_data',
-            'matrix_lendingclub', 'matrix_set_lendingclub', 'matrix_operations']
+            'matrix_lendingclub', 'matrix_set_lendingclub', 'matrix_operations', 'matrix_set_operations']
 
 if __name__ == '__main__':