+++ transitionMatrix/estimators/simple_estimator.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 
-# (c) 2017-2018 Open Risk, all rights reserved
+# (c) 2017-2019 Open Risk, all rights reserved
 #
 # TransitionMatrix is licensed under the Apache 2.0 license a copy of which is included
 # in the source distribution of TransitionMatrix. This is notwithstanding any licenses of