+++ transitionMatrix/model.py
@@ -1,6 +1,6 @@
 # encoding: utf-8
 
-# (c) 2017-2018 Open Risk (https://www.openriskmanagement.com)
+# (c) 2017-2019 Open Risk (https://www.openriskmanagement.com)
 #
 # TransitionMatrix is licensed under the Apache 2.0 license a copy of which is included
 # in the source distribution of TransitionMatrix. This is notwithstanding any licenses of
@@ -14,6 +14,7 @@
 
 """ This module provides the key transition matrix objects
 
+* CreditCurve_ implements the functionality of a collection of credit (default curves)
 * TransitionMatrix_ implements the functionality of single period transition matrix
 * TransitionMatrixSet_ provides a container for a multiperiod transition matrix collection
 * StateSpace holds information about the stochastic system state space
@@ -22,18 +23,168 @@
 """
 
 import json
-import os
-
 import numpy as np
+import os
 import pandas as pd
+import transitionMatrix as tm
 from scipy.linalg import logm
 
-import transitionMatrix as tm
+
+class CreditCurve(np.matrix):
+    """ The _`CreditCurve` object implements a typical collection of `credit curves <https://www.openriskmanual.org/wiki/Credit_Curve>`_.
+    The class inherits from numpy matrices and implements additional properties specific to curves.
+
+    """
+
+    def __new__(cls, values=None, json_file=None, csv_file=None):
+        """ Create a new credit curve set. Different options for initialization are:
+
+        * providing values as a list of list
+        * providing values as a numpy array  (The rows are the different curves, the columns are different periods)
+        * loading from a csv file
+        * loading from a json file
+
+        Without data, a default identity matrix is generated with user specified dimension
+
+        :param values: initialization values
+        :param json_file: a json file containing transition matrix data
+        :param csv_file: a csv file containing transition matrix data
+        :type values: list of lists or numpy array
+        :returns: returns a CreditCurve object
+        :rtype: object
+
+        .. note:: The initialization in itself does not validate if the provided values form indeed a credit curve
+
+        :Example:
+
+        .. code-block:: python
+
+            A = tm.CreditCurve(values=[[0.1, 0.2, 0.3], [0.2, 0.6, 0.8], [0.01, 0.02, 0.06]])
+
+        """
+        obj = None
+        if values is not None:
+            # Initialize with given values
+            obj = np.asarray(values).view(cls)
+        elif json_file is not None:
+            # Initialize from file in json format
+            q = pd.read_json(json_file)
+            obj = np.asarray(q.values).view(cls)
+        elif csv_file is not None:
+            # Initialize from file in csv format
+            q = pd.read_csv(csv_file, index_col=None)
+            obj = np.asarray(q.values).view(cls)
+        # validation flag is set to False at initialization
+        obj.validated = False
+        # temporary dimension assignment (must validated for squareness)
+        obj.dimension = obj.shape[0]
+        return obj
+
+    def to_json(self, file):
+        """
+        Write credit curves to file in json format
+
+        :param file: json filename
+        """
+
+        q = pd.DataFrame(self)
+        q.to_json(file, orient='values')
+
+    def to_csv(self, file):
+        """
+        Write credit curves to file in csv format
+
+        :param file: csv filename
+        """
+
+        q = pd.DataFrame(self)
+        q.to_csv(file, index=None)
+
+    def to_html(self, file=None):
+        html_table = pd.DataFrame(self).to_html()
+        if file is not None:
+            file = open(file, 'w')
+            file.write(html_table)
+            file.close()
+        return html_table
+
+    def validate(self, accuracy=1e-3):
+        """ Validate required properties of a credit curve set. The following are checked
+
+        1. check that all values are probabilities (between 0 and 1)
+        2. check that values are non-decreasing
+
+        :param accuracy: accuracy level to use for validation
+        :type accuracy: float
+
+        :returns: List of tuples with validation messages
+        """
+        validation_messages = []
+
+        curve_set = self
+        curve_set_size = curve_set.shape[0]
+        curve_set_periods = curve_set.shape[1]
+
+        # checking that values of curve_set are within allowed range
+        for i in range(curve_set_size):
+            for j in range(curve_set_periods):
+                if curve_set[i, j] < 0:
+                    validation_messages.append(("Negative Probabilities: ", (i, j, curve_set[i, j])))
+                if curve_set[i, j] > 1:
+                    validation_messages.append(("Probabilities Larger than 1: ", (i, j, curve_set[i, j])))
+        # checking monotonicity
+        for i in range(curve_set_size):
+            for j in range(1, curve_set_periods):
+                if curve_set[i, j] < curve_set[i, j - 1]:
+                    validation_messages.append(("Curve not monotonic: ", (i, j)))
+
+        if len(validation_messages) == 0:
+            self.validated = True
+            return self.validated
+        else:
+            self.validated = False
+            return validation_messages
+
+    def hazard_curve(self):
+        """
+        .. Todo:: Compute hazard rates
+        :return:
+        """
+        pass
+
+    def characterize(self):
+        """ Analyse or classify a credit curve according to its properties
+
+        * slope of hazard rate
+
+        .. Todo:: Further characterization
+        """
+
+        pass
+
+    def print(self, format_type='Standard', accuracy=2):
+        """ Pretty print a set of credit curves
+
+        :param format_type: formatting options (Standard, Percent)
+        :type format_type: str
+        :param accuracy: number of decimals to display
+        :type accuracy: int
+
+        """
+        for s_in in range(self.shape[0]):
+            for s_out in range(self.shape[1]):
+                if format_type is 'Standard':
+                    format_string = "{0:." + str(accuracy) + "f}"
+                    print(format_string.format(self[s_in, s_out]) + ' ', end='')
+                elif format_type is 'Percent':
+                    print("{0:.2f}%".format(100 * self[s_in, s_out]) + ' ', end='')
+            print('')
+        print('')
 
 
 class TransitionMatrix(np.matrix):
     """ The _`TransitionMatrix` object implements a typical (one period) `transition matrix <https://www.openriskmanual.org/wiki/Transition_Matrix>`_.
-    The classs inherits from numpy matrices and implements additional properties specific transition matrices. It form the building block of the
+    The class inherits from numpy matrices and implements additional properties specific transition matrices. It forms the building block of the
     TransitionMatrixSet_ which holds a collection of matrices in increasing temporal order
 
     """
@@ -331,6 +482,7 @@ class TransitionMatrixSet(object):
 
         """
 
+        self.dimension = dimension
         if values is not None:
             # Copy a single matrix to all periods
             if method is 'Copy':
@@ -557,7 +709,7 @@ class TransitionMatrixSet(object):
 
         # TODO Make absorbing state an attribute of Matrix and MatrixSet
         # Default state hardwired to be highest matrix element
-        Default = self.entries[0].dimension - 1
+        Default = self.dimension - 1
         Periods = len(self.periods)
 
         iPD = np.zeros(Periods)
@@ -576,6 +728,21 @@ class TransitionMatrixSet(object):
         elif self.temporal_type is 'Incremental':
             pass
         return iPD, cPD, hR, sR
+
+    def default_curve_set(self):
+        """ Calculate the cumulative probabilities (credit curves) for all ratings
+
+        """
+        Default = self.dimension - 1
+        values = []
+        for c in range(self.dimension - 1):
+            curve = []
+            for k in range(len(self.periods)):
+                curve.append(self.entries[k][c, Default])
+            values.append(curve)
+
+        credit_curves = CreditCurve(values=values)
+        return credit_curves
 
 
 class StateSpace(object):